# res/openapi/huki-ebook-openapi.generated.json
{
  "components": {
    "schemas": {
...
      "promotion_CreateVoucherDto": {
        "properties": {
          "code": {
            "example": "HUKI10",
            "type": "string"
          },
          "description": {
            "example": "Áp dụng cho đơn hàng từ 100.000đ",
            "type": "string"
          },
          "expiresAt": {
            "example": "2026-08-31T23:59:59Z",
            "type": "string"
          },
          "maxDiscountAmount": {
            "description": "Maximum discount amount",
            "example": 50000,
            "type": "number"
          },
          "maxUsagePerUser": {
            "description": "Max usage per user",
            "example": 1,
            "type": "number"
          },
          "minOrderAmount": {
            "description": "Minimum order amount",
            "example": 100000,
            "type": "number"
          },
          "name": {
            "example": "Giảm 10% cho đơn hàng đầu tiên",
            "type": "string"
          },
          "scope": {
            "default": "PLATFORM",
            "enum": [
              "PLATFORM",
              "STORE"
            ],
            "type": "string"
          },
          "startsAt": {
            "example": "2026-08-01T00:00:00Z",
            "type": "string"
          },
          "storeId": {
            "description": "Store ID if scope is STORE",
            "type": "string"
          },
          "totalUsage": {
            "description": "Total usage limit",
            "example": 100,
            "type": "number"
          },
          "type": {
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT",
              "FREE_SHIPPING"
            ],
            "example": "PERCENTAGE",
            "type": "string"
          },
          "value": {
            "description": "Percentage (1-100) or fixed amount",
            "example": 10,
            "type": "number"
          }
        },
        "required": [
          "code",
          "name",
          "type",
          "value",
          "scope",
          "startsAt",
          "expiresAt"
        ],
        "type": "object"
      },
...
      "promotion_UpdateVoucherDto": {
        "properties": {
          "description": {
            "type": "string"
          },
          "maxDiscountAmount": {
            "type": "number"
          },
          "maxUsagePerUser": {
            "type": "number"
          },
          "minOrderAmount": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "scope": {
            "enum": [
              "PLATFORM",
              "STORE"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "EXPIRED",
              "USED_UP"
            ],
            "type": "string"
          },
          "totalUsage": {
            "type": "number"
          },
          "type": {
            "enum": [
              "PERCENTAGE",
              "FIXED_AMOUNT",
              "FREE_SHIPPING"
            ],
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "promotion_ValidateVoucherDto": {
        "properties": {
          "code": {
            "example": "HUKI10",
            "type": "string"
          },
          "orderSubtotal": {
            "description": "Order subtotal",
            "example": 200000,
            "type": "number"
          },
          "storeId": {
            "description": "Store ID for store-scoped vouchers",
            "type": "string"
          }
        },
        "required": [
          "code",
          "orderSubtotal"
        ],
        "type": "object"
      },
      "shipping_AssignDeliveryStaffDto": {
        "properties": {
          "staffId": {
            "type": "string"
          }
        },
        "required": [
          "staffId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "commerce_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "community_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "identity_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-auth": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "JWT-refresh": {
        "bearerFormat": "JWT",
        "description": "Enter JWT token",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      },
      "shipping_bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "https://huki.ebook.com",
      "name": "HUKI Support",
      "url": "support@huki.ebook.com"
    },
    "description": "\n# HUKI EBOOK Platform API\n\nWelcome to the HUKI EBOOK API documentation. This API provides endpoints for managing books, orders, payments, and more.\n\n## Base URL\n`http://localhost:3000/api/v1`\n\n## Authentication\nMost endpoints require JWT Bearer token authentication.\n1. Register/Login to get access token\n2. Include header: `Authorization: Bearer <token>`\n\n## Services\n| Service | Port | Description |\n|---------|------|-------------|\n| Identity | 3001 | Auth, Users, Sessions |\n| Business | 3002 | Business, Stores, Members |\n| Commerce | 3003 | Books, Cart, Orders, Payments |\n| Shipping | 3004 | Shipments, Addresses |\n| Community | 3005 | Forum, Chat, Reviews |\n| Promotion | 3007 | Vouchers, Banners |\n\n## Common Headers\n| Header | Required | Description |\n|--------|----------|-------------|\n| Content-Type | Yes | application/json |\n| Authorization | No | Bearer token |\n| Idempotency-Key | For checkout | Unique key for idempotent operations |\n\n## Error Codes\n- 400: Bad Request\n- 401: Unauthorized\n- 403: Forbidden\n- 404: Not Found\n- 409: Conflict\n- 500: Server Error\n\n---\n*Last updated: 2026-08-24*\n    ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://huki.ebook.com/terms",
...
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/internal/vouchers/validate": {
      "post": {
        "operationId": "InternalPromotionsController_validateVoucher",
        "parameters": [
          {
            "in": "header",
            "name": "x-internal-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Validate voucher for checkout (internal)",
        "tags": [
          "Internal"
        ]
      }
    },
...
    "/api/v1/shipments": {
      "get": {
        "operationId": "ShipmentsController_list",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "PENDING",
                "PICKED_UP",
                "IN_TRANSIT",
                "OUT_FOR_DELIVERY",
                "DELIVERED",
                "RETURNED",
                "CANCELLED",
                "FAILED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated shipment list"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List shipments visible to the authenticated actor",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}": {
      "get": {
        "operationId": "ShipmentsController_detail",
        "parameters": [
          {
            "description": "Shipment UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shipment details"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Shipment is outside the actor scope"
          },
          "404": {
            "description": "Shipment not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get shipment details",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}/assign": {
      "post": {
        "operationId": "ShipmentAssignmentController_assign",
        "parameters": [
          {
            "description": "Shipment UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/shipping_AssignDeliveryStaffDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Delivery staff assigned"
          },
          "400": {
            "description": "Shipment cannot be assigned"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Platform administrator role required"
          },
          "404": {
            "description": "Shipment or delivery staff not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign delivery staff to a shipment (platform admin)",
        "tags": [
          "Shipment assignment"
        ]
      }
    },
...
    "/api/v1/vouchers": {
      "get": {
        "description": "Returns a paginated list of vouchers.",
        "operationId": "VouchersController_findAll",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "EXPIRED",
                "USED_UP"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "enum": [
                "PLATFORM",
                "STORE"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "storeId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of vouchers"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all vouchers",
        "tags": [
          "Vouchers"
        ]
      },
      "post": {
        "description": "Creates a new voucher. Requires PLATFORM_ADMIN role.",
        "operationId": "VouchersController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/promotion_CreateVoucherDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Voucher created successfully"
          },
          "400": {
            "description": "Invalid voucher data"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Requires PLATFORM_ADMIN role"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new voucher",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/api/v1/vouchers/{id}": {
      "delete": {
        "description": "Deletes a voucher. Requires PLATFORM_ADMIN role.",
        "operationId": "VouchersController_delete",
        "parameters": [
          {
            "description": "Voucher ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher deleted successfully"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Requires PLATFORM_ADMIN role"
          },
          "404": {
            "description": "Voucher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete voucher",
        "tags": [
          "Vouchers"
        ]
      },
      "get": {
        "description": "Returns voucher details by ID.",
        "operationId": "VouchersController_findOne",
        "parameters": [
          {
            "description": "Voucher ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher details"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Voucher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get voucher by ID",
        "tags": [
          "Vouchers"
        ]
      },
      "patch": {
        "description": "Updates voucher details. Requires PLATFORM_ADMIN role.",
        "operationId": "VouchersController_update",
        "parameters": [
          {
            "description": "Voucher ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/promotion_UpdateVoucherDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Voucher updated successfully"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "403": {
            "description": "Requires PLATFORM_ADMIN role"
          },
          "404": {
            "description": "Voucher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update voucher",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/api/v1/vouchers/code/{code}": {
      "get": {
        "description": "Returns voucher details by promo code.",
        "operationId": "VouchersController_findByCode",
        "parameters": [
          {
            "description": "Voucher code",
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher details"
          },
          "401": {
            "description": "Invalid or missing token"
          },
          "404": {
            "description": "Voucher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get voucher by code",
        "tags": [
          "Vouchers"
        ]
      }
    },
    "/api/v1/vouchers/validate": {
      "post": {
        "description": "Validates a voucher code for use with an order.",
        "operationId": "VouchersController_validate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/promotion_ValidateVoucherDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validation result"
          },
          "400": {
            "description": "Invalid voucher code or conditions not met"
          },
          "401": {
            "description": "Invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validate voucher",
        "tags": [
          "Vouchers"
        ]
      }
    }
  },
  "servers": [
...
}